HBX-1675: Move class 'org.hibernate.tool.hbmlint.detector.SchemaByMetaDataDetector' to 'org.hibernate.tool.internal.export.lint.SchemaByMetaDataDetector'
